+++ src/kit/Utils.hh
@@ -12,6 +12,7 @@ protected:
 public:
     Touch *touch;
     Pad *pad;
+    I18n *i18n;
     Utils();
     virtual void read();
 };