+++ src/kit/ui/UiTexts.cpp
@@ -1,6 +1,6 @@
 #include "UiTexts.hpp"
 
-#include <stdarg.h>
+using namespace ufal::unilib;
 
 UiTexts::UiTexts() {
     family = "Roboto";
@@ -120,13 +120,31 @@ TextData UiTexts::getTextData(std::string text, TextStyle textStyle, bool italic
     return textData;
 }
 
-//private
-
 //set text to uppercase
 std::string UiTexts::toUppercase(std::string text) {
-    std::transform(text.begin(), text.end(),text.begin(), ::toupper);
+
+    utf8::decode(text, text32);
+
+    for (auto&& chr :  text32) chr = unicode::uppercase(chr);
+
+    utf8::encode(text32, text);
+
     return text;
 }
+
+//set text to lowercase
+std::string UiTexts::toLowercase(std::string text) {
+
+    utf8::decode(text, text32);
+
+    for (auto&& chr :  text32) chr = unicode::lowercase(chr);
+
+    utf8::encode(text32, text);
+
+    return text;
+}
+
+//private
 
 void UiTexts::calcTextData(std::string text, TextStyle textStyle, bool italic) {