+++ src/kit/utils/IME.hpp
@@ -39,7 +39,7 @@ private:
 
 	public:
 		IME();
-		std::string getUserText(const char *title, const char *showText = "", unsigned int type = SCE_IME_TYPE_BASIC_LATIN, unsigned int option = 0);
+		std::string getUserText(const char *title, const char *showText = "", unsigned int type = SCE_IME_TYPE_BASIC_LATIN, int maxTextLength = 128, unsigned int option = 0);
 		SceCommonDialogStatus getStatus() const;
 };