+++ src/views/I18nSample.cpp
@@ -6,9 +6,9 @@ I18nSample::I18nSample(const char *name) : View(name) {
 
 void I18nSample::contents() {
 
-    //ui->texts->draw(10, 10, Body1, THEME_PRIMARY, utils->i18n->getI18nByCat("test")["phrase"]);
-    ui->texts->draw(10, 60, Body1, THEME_PRIMARY, frLang->getI18nByCat("test")["phrase"]);
-    ui->texts->draw(10, 110, Body1, THEME_PRIMARY, enLang->getI18nByCat("test")["phrase"]);
+    ui->texts->draw(10, 10, Body1, THEME_PRIMARY, defaultL["phrase"]);
+    ui->texts->draw(10, 60, Body1, THEME_PRIMARY, frL["phrase"]);
+    ui->texts->draw(10, 110, Body1, THEME_PRIMARY, enL["phrase"]);
 
     back = ui->buttons->containedDraw("Back", 800, 480);
 }
@@ -17,5 +17,14 @@ void I18nSample::controls() {
     if (ui->buttons->onTouch(back, utils->touch->lastClickPoint)) {
         viewsController->setActualView("Welcome");
     }
+}
+
+void I18nSample::mounted() {
+    I18n *frLang = new I18n(SCE_SYSTEM_PARAM_LANG_FRENCH);
+    I18n *enLang = new I18n(SCE_SYSTEM_PARAM_LANG_ENGLISH_US);
+
+    defaultL = utils->i18n->getI18nByCat("test");
+    frL = frLang->getI18nByCat("test");
+    enL = enLang->getI18nByCat("test");
 }