+++ src/views/I18nSample.hh
@@ -6,12 +6,16 @@
 class I18nSample: public View {
 private:
     ZoneEvent back;
-    I18n *frLang = new I18n(SCE_SYSTEM_PARAM_LANG_FRENCH);
-    I18n *enLang = new I18n(SCE_SYSTEM_PARAM_LANG_ENGLISH_US);
+    I18n *frLang;
+    I18n *enLang;
+    std::map<std::string, std::string> defaultL, frL, enL;
+
 public:
     I18nSample(const char *name);
     void contents() override;
     void controls() override;
+
+    void mounted() override;
 };
 
 #endif //VITAMATERIALKIT_I18NSAMPLE_HH